requirements.txt
fastapi
jinja2
praw
boto3
python-dotenv
uvicorn